requirements.txt
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorlover==0.3.0
cufflinks==0.14.6
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.7
entrypoints==0.3
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
plotly==3.6.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
pyzmq==18.0.0
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
